Implemented the styling for the question card component
