+++ bandersnatch-react/public/index.html
@@ -8,9 +8,10 @@
       name="description"
       content="Web site created using create-react-app"
     />
+    <link href="https://unpkg.com/tailwindcss@^2/dist/tailwind.min.css" rel="stylesheet">
     <title>Bandersnatch Game</title>
   </head>
-  <body>
+  <body class="bg-pink-300 flex justify-center items-center h-screen ">
     <div id="root"></div>
   </body>
 </html>
